Separated docker. Updated handlers and functions.

+++ gw-exchanger/internal/storages/model.go
@@ -3,5 +3,5 @@ package storages
 type ExchangeRate struct {
 	FromCurrency string
 	ToCurrency   string
-	Rate         float64
+	Rate         float32
 }
