+++ gw-exchanger/internal/storages/storage.go
@@ -2,5 +2,5 @@ package storages
 
 type Storage interface {
 	GetExchangeRates() (map[string]float64, error)
-	GetExchangeRate(fromCurrency, toCurrency string) (float64, error)
+	GetExchangeRate(fromCurrency, toCurrency string) (float32, error)
 }
